// src/app/service/model/book.ts
/**
 * BookStore APIs
 * BookStore APIs exposed from a Java EE back-end to an Angular front-end
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ahmedibraheem10@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Book resource representation
 */
export interface Book { 
    /**
     * Identifier
     */
    id?: number;
    isbn: string;
    /**
     * Title of the book
     */
    title: string;
    imageUrl?: string;
    language?: string;
    unitCost?: number;
    nbOfPages?: number;
    publicationDate?: Date;
    /**
     * Description of the book
     */
    description?: string;
}